Revert Stream.Cons to the 2.12 encoding
- The `lazy val` required an extra bitmap field which increases the
  instance size from 24 to 32 bytes.
- Capturing `tl` in the initialization code prevented it from being
  garbage collected as soon as the tail has been computed and the
  initializer is no longer needed.

Fixes https://github.com/scala/bug/issues/11696, at least in theory.
I was unable to reproduce the failure with the code from the original
bug report.

LazyList is not affected. It already releases the initializer after
running it, and the instance size (on x64) is 24 bytes with or without
the bitmap field for the lazy val.