Limit early initializer deprecation to -Xsource:2.14.
Previously, a deprecation message was always reported for early
initializers with the following message:

```
early initializers are deprecated; they will be replaced by trait parameters in 2.14, see the migration guide on avoiding var/val in traits.
```

This commit changes the deprecation message so that it's only reported
when users have enabled `-Xsource:2.14`, which adds support for trait
parameters.