Avoid ObjectRefs in two hot paths in compiler
Avoid capturing synthetic patmat vars in closures, etc, which
gives rise to an ObjectRef to box the var. There is currently a
bug in the pattern matcher that emits these more of these vars
under `-optimize` -- the logic was not updated woth `-opt` was
introduced in 2.12.

Rewrite `RefChecks.transform` in terms of vals to avoid another
ObjectRef.