Refactor Javap support to support JDK 9
Repair broken test, command options changed.

The javap tool can be run a few ways from Javap.

For JDK 8, add tools.jar to the class loader and
load the internal Task reflectively.

Otherwise, if we seem to be loaded by the App loader,
so that we can access platform classes, also just
load the internal Task. There is no tools.jar.

Another strategy is to use the ToolProvider interface,
which has the drawback that the REPL output directory
must be a real directory on the class path.

If we are loaded by the bootstrap loader, trying to
load the platform classes will fail. Since the REPL
class loader is a child of the platform loader,
we could use its compiler to compile the task class that
can run the internal tool non-reflectively.

Remove that adapter option that compiled an adapter class
on the REPL class path. Instead, recommend running
`-nobootcp` or `-Yrepl-outdir d -toolcp d`. The first
option makes javap visible to REPL; the second option
makes REPL output visible to the ToolProvider-provided tool.