Stricten @varargs check.
If you get it wrong, the JVM will complain. Now we complain first.

Not a refcheck for some reason.

I hope it's ok to do a "good first issue" as a "good first issue after
several months' hiatus". For what it's worth, it was. Who doesn't like
adding new compiler errors?

Closes scala/bug#11714.