Don't warn about auto-application for nullary methods overriding nilary
Let's call

    def f: Int   // nullary
    def f(): Int // nilary

When overriding a nilary method with a nullary one, the compiler
converts the override in Namers and adds the `()` parameter list.

    trait T { def f(): Int }
    class C extends T { def f: Int = 1 } // converted to `def f(): Int`

Calling `(c: C).f` makes the compiler insert `()`, and warn about that
fact under `-Xsource:2.14`.

This PR adds an annotation to `C.f` to remember that the method was
originally nullary. I initially tried with a symbol attachment, but that
doesn't survive pickling, so doesn't work for separate compilation. The
annotation is just a WIP hack, we'd need to hide the secret handshake
better IRL.

The adaptation still leaks though. For example:

    trait T { def f(): Int }
    class A           { def f: Int = 0 }
    class B extends T { def f: Int = 0 }
    (new A).f() // error: Int does not take parameters
    (new B).f() // ok

Maybe, instead of adding more hacks to selectively undo the adaptation,
it would be better to teach the compiler about the fact that `f` may
override `f()`.