Add specialised SubstTypeMap to replace with wildcards.
We introduce two middle "abstract" classes, AbstractSubstTypeMap and
AbstractSubstMap, and we keep as much of the logic as possible in
there, using the "template method" design pattern.
In this case, we make the logic dependent upon a "find" method,
which looks for a symbol (key) in the "map", which in the common case
is implemented as two lists.

We then introduce a special class `SubstByWildcardTypeMap`, which
does not have a "to" list, but always returns WildcardType.
This allows us to replace the calls to
`new SubstTypeMap(xs, xs map (_ => WildcardType))`,
with calls to `new SubstByWildcardTypeMap(xs)`,
thus avoiding the cost of creating that bogus list of values.