Type Comparers: reuse the `SubstSymMap` instances.
In the TypeComparers trait, there are a couple of "corresponds"
loops that call in multiple times to the `Type.substSym` method.
But this method creates a new SubstSymMap object upon every call.
This is creating a notable number of allocations of SubstSymMap.

To avoid these allocations, we extract that map as a local variable.