Scope: use direct iterator to checkNoAbstractMembers
We introduce a simple iterator for the Scope class that, unlike the
normal iterator, processes the elements as reachable through the
entries, and thus needs no extra list (stack) to LIFO its elements.
As long as what we want to do with the iterator is a forall or exists,
as is the case in `checkNoAbstractMethods`, the order matters not.

The motivation was that we noticed `checkNoAbstractMembers` to be a
source of allocations of `::` objects, due to use of `Scope.toList`
from `Scope.partition`.