Browse safely after erasure
Kinda like those virus-scanning toolbars used to advertise.

Probably no one but me and Dale (who asked for it lo so many years back)
will notice this, but let's do it anyways. The opportunity for a "why
browse?" pun's been taken away from me, but the feature does get some use.

I think this is the only missing type that could plausibly be on a tree
at the end of a phase.

Fixes scala/bug#9774.