ByteCode Body Builder: avoid middle lists allocations.
Within these methods, a set of intermediate linked lists are created,
to represent sequences of arguments passed to the ASM api.
However, since ASM is a Java library, these sequences of arguments
need to be passed as arrays, so the lists are converted.
This means that the List objects are transient and redundant.

The changes in this commit directly create the arrays passed to the
ASM methods, and fill them with the elements in the right order.