Reuse StringBuilder used for printing BTypes.
In the Bytecode backend, to print the BType to a String (which we need
to get the `descriptor` in ASM) we use a `StringBuilder`.
Unlike builders in the Scala collections, a StringBuilder does not
share memory with the final object: a call to `toString` dumps all its
content into a new String, backed by a new array.  So, using a new
StringBuilder every time means a lot of single-use memory allocation.

To prevent this, we extract the toString printing logic to an object,
and use a single StringBuilder that we recycle for all calls.
Before and after a printing starts, we clear the builder.

However, because of the `-Ybackend-parallelism` setting, we could
have more several threads using that StringBuilder. Since we do not
need sharing, we can get thread safety using a ThreadLocal variable.