Avoid Option.WithFilter allocations.
When we use a for-comprehension (a foreach in this case) over the
Option class, and when we use an `if` clause in the comprehension
(which corresponds to using a `withFilter`), then an object of the
class `Option.withFilter` is created.

Checking the same predicate inside the body of the for comprehension
should avoid that allocation.