SourceFile: avoid using unboxed integers.
We noted that the computation of `lineIndices` was a source of
allocations of boxed Integer objects. The reason is that, by using an
`ArrayBuffer` (which is backed by an array of objects), it needs
to box those numbers, allocating integers. To avoid this:

- We modify the code to directly compute the resulting `Array[Int]`,
  and directly write into the
- We replace the use of ranges and for-iterations with while loops.

One downside is that we now have to run two passes through the
`cs` array of characters: one to count how many end-of-lines we have,
and another to actually fill those positions.