nReuse arrays in PickleBuffer and related.
The class PickleBuffer, and its subclass Scan, are a small allocation
hotspot, of int[], and Object[] instances.
this is because of the "index" object that is created in each
Picklebuffer, which is used to indicate special stop points in the
data given in the `_bytes` array.

The goal of this commit is to avoid allocations in there by reusing
and recycling a mutable instance, one in which we try to keep and
clear and reuse the arrays that are created.