The compiler bridge compiles
Adjust the packages

Split up the classes pulled in from Zinc, the internals (phases,
global subclass, reporter) stay in scala.tools.nsc.incremental, the
compiler bridge implementation classes (CompilerInterface,
ScaladocInterface, ConsoleInterface) go to the compiler-bridge
subproject, in the xsbt package.

The bridge implementation cannot be in the compiler subproject, because
the code depends on repl, scaladoc, interacitve, which are their own
subprojects.

We can still decide to put the compiler interface in scala-compiler.jar,
like we do for repl / scaladoc, or keep it as a separate jar.