Make compiler Scope doubly linked
This avoids the need to materialize a temporary list to iterate
through the elements in insertion order.

The internally cached `List[Symbol]` is removed, and operations
are all now based on iterators.

The iterators will fail fast when a scope is mutated during iteration.
So far, I've found a few use sites that needed refactoring to avoid
this problem.

The previous definition of `size` was inconsistent with
`itertaor.size`, as the latter only considered elements owned
by the current scope (ie, not ones from outer scopes in1 which
this scope is nested with `newNestedScope`). After reviewing its
usage in the compiler, I've redefined it only consider direct elements.