Remove NameOps implicit class (syntax extension), use object
The NameOps implicit class is used to provide some extension
methods to the Name classes and subclasses. However, because
of some problem of outer references (being defined in a cake
slice) they could not be turned into a plain `AnyVal`, so
there was an allocation cost to them.

While that allocation cost was not significant (0.2%), it is
easily avoidable, with some syntactic salt and vinegar.