Get rid of the last implicit conversions from String to Names.
This requires a fair amount of explicit calls to `createNameType`
in `StdNames`, which is not pleasing to the eye, but the result is
a saner codebase nevertheless.

We use the opportunity to replace the `n"foo"` interopolator trick,
which was used to avoid spurious "possible missing interpolator"
warnings, by a simple concatenation like `"$" + "foo"`.

In the `scala-reflect` API, we do not remove the methods to
preserve binary compatibility. We simply drop their `implicit`
marker, which is not source compatible but is binary compatible.