Plug many variance holes (pos and neg)
The variance checks root in `RefChecks` for `TypeTree`s is modified
to check only the variance of `PolyType`s (this includes type lambdas).
Checking the definition of the `PolyType`s' owners is not correct,
because `relativeVariance` doesn't work for types in arbitrary position.

To check the variance of `PolyType`s we introduce a new type map:
`PolyTypeVarianceMap`. This is similar to the way variance is checked
for type lambdas in dotty, where type lambdas are encoded explicitly.

We also add explicit tracking for lower bounds of `PolyType`s.
For a type parameter of a `PolyType`, the variance should not be
flipped (or equivalently, flipped twice) in the lower bound.
This applies to both `ValidateVarianceMap` and `PolyTypeVarianceMap`.
It fixes variance bugs for types appering in the bounds of HKTs.

Finally, we don't skip local definitions entirely (this is unsound),
but instead cutoff `relativeVariance` at local boundaries.