SI-6623 -Yrepl-use-magic-imports avoids nesting $iw wrappers
Rather than adding a wrapper object for each import in the session
history, just use a single wrapper preceded by the imports which
have been interspersed with a magic import to bump context depth.

Code is still ordinarily wrapped in a `$read` object.

This is a step toward 6623-like transparency.

`retronym` takes the blame for this innovation.
`adriaanm` collaborated in its commission.
`somsnytt` batted clean-up.