Lint confusing by-name conversion of block result
When adapting the result expression of a block with
an implicit conversion that takes a by-name parameter,
only the result expression is adapted; one may incorrectly
think the whole block (as an arg) receives by-name
evaluation semantics.

This lint warns when it notices an implicit applied
in result position of a multiline block.