Bug fixes in divergent implicits checker
We can't rely on type equality (=:=) to indicate domination
or covering when the types contain wildcards (?) because
wildcards are equal to anything by definition (e.g. ? =:= (?, ?)).
But we explicitly add those wildcards in the stripped method.
Since type equality is not essential to the computation we omit it,
unless the complexity is exactly equal, then we need to check.