Literals can form part of stable paths
References to enum constants or (literal-typed) class types get replaced
with `Literal` trees, which can be selected from and should be
considered stable for the purpose of the import check.

The second and third tests in the enclosed partest case argue for this
change (as does the fact that the test is accepted by Dotty): removing
`final` from the vals makes them compile prior to this change.

Fixes scala/bug#8855