Preserve position of argument of right-associative method
To preserve evaluation order, `x :: y` is rewritten not to `y.::(x)` but
to `{ val x$1 = x; y.::(x$1) }`. But, the `Ident(x$1)` tree is the
target of any type-mismatch errors issued later on, so it should steal
the position of `x` for the error message.

The important effect of this important change:
```
% diff /code/scala/test/files/neg/t4701-neg.log /code/scala/test/files/neg/t4701.check--- t4701-neg.log
+++ t4701.check
@@ -4,3 +4,3 @@
   hasType[HL[String]](nnn :: HN) // type mismatch error should have position at `nnn`
-                          ^
+                      ^
```

Fixes scala/bug#4701