Sort inline request before selecting according to method size limit
`selectRequestsForMethodSize` filters a list of inline requests for a
method to make sure the final size of the method is within a defined
limit.

We need to make sure to sort the list of requests before calling that
method, otherwise the final bytecode is non-deterministic. The method
list comes from `callGraph.callsites`, which is a
`Map[MethodNode, Map[MethodInsnNode, Callsite]]`, so non defined order.