Warn about unintended underscores
If a pattern definition binds no variables, it is
probably a mistake if it is marked implicit,
because it introduces no implicit values,
or if it is a template statement, because
it accidentally introduces a template member.

Also since the Scouts are selling cookies this
week, I did some Girl Scouting and tweaked doc
for partest and improved some tests, including
a test that fails under Java 9 because it
uses underscore.