Make -Yrepl-class-based set -Yrepl-use-magic-imports
The Spark REPL uses -Yrepl-class-based and the change around
ImportHandler in interpreter/Imports (and that introduced the custom
wrapperCleanup phase for the REPL) requires -Yrepl-use-magic-imports be
set in order for the Spark REPL to continue to work correctly.  Rather
than affect all existing Spark releases that choose to upgrade Scala, we
preserve the functioning of the Spark REPL by enabling "magic imports"
when using the class-based wrappers.