[backport] Align package detection in Java with Java
Java packages are absolute, not relative.

So for a package symbol to be a typed identifier,
it must be effectively `_root_`ed.

I don't know why people like to call their
packages `java`, `scala`, and `util`, but
the conflict happens for any package name
appearing twice in a dotted name and
accessed from the bowels.

Also bestowed upon partest the ability to
descend a directory, which is technology
available only since Java 7. This was the
reason they forked Vulpix/Xipluv.

Noticed that a couple of other tests seem
to be mistakenly nested.