Make REPL dead backreference elimination work with nested wrapped
The custom phase that does this needs to descend deeper than
the first `Template`.

It now works bottom up (from the innermost `$iw`) and runs the
unused analysis at each level, feeding the possibly simpilfied
trees into the analysis of the enclosing level.

I also needed to add a special case for the `lineXY$read`
variable names which should be considered private for the
purpose of this analysis.