Rework insertion of stabilizing definitions
Introduce a typer mode to know when we're type checking a qualifier
of a select, or a function part of an application. When a stabilizing
qualifier is created, insert it once we get back out of this mode.

In `x.y.foo(bar)`, if `foo` has a second implicit argument list, make
sure that the block with the stabilizer is only created after the
implicit is inferred, i.e., run `adapt` on `$stabilizer.foo(bar)`
to get `$stabilizer.foo(bar)(<implicit arg>)` first, then create

```
{
  val $stabilizer = x.y
  $stabilizer.foo(bar)(<implicit arg>)
}
```

Before, we would create

```
{
  val $stabilizer = x.y
  $stabilizer.foo(bar)
}
```

then assign a `MethodType` to the block, and run `adapt`. This worked in
some situations by chance, but not always.