Mutable Symbol substitution to `cloneSymbols` and `copyRefinedType
We found allocation hotspots of `SubstSymMap` objects in methods
`cloneSymbols`, and `copyRefinedType`. To cool them, we introduce
Reusable Instances with mutable symbol substitution map.

To introduce this mutable symbol substitutions, we use two middle
"abstract" classes, `AbstractSubstTypeMap` and `AbstractSubstMap`,
that keep as much of the logic as possible in this template.
The logic depends on a "find" method, which looks for a symbol (key)
in the "map", which in the common caseb is implemented as two lists.