Fixes bug in scala docs while navigating the page with space bar/ shift+space
space/shift+space in all modern browsers scroll down/up respecitively.
But this functionality doesn't work on the scala docs. Because of
combination of  z-index and position of the elements in the dom.

Detailed Diagnosis:
- search bar at the top of the page has a zindex of 103 And, the rest of
the page has the z-index of 100. By making the page position: abolute.
The effective dom size became the  size of search bar (as only the
search bar is in the focus).

The fix is to make the rest of the page position:relative (default), so
the dom size is calculated correctly.

Tested in chrome, edge and firefox.