Warn constant promotions
Warn if widening of literal Int or Long is not safe,
which is defined as the conversion roundtrips, the
test used for harmonization.

The test means that if I want to convert precisely back to my
original value, I can do so. It does not mean the converted
value maintains accuracy or number of significant digits.

An attachment is used to defer the warning, since typechecking
may consider expressions which are discarded, such as
conversions during implicit search.