Use Java semantics for imports in .java files
An static member of a superclass may be referred to in an
static import in Java sources qualified by a reference to the subclass.

We've accounted for related problems previously in `typedIdent`
and `typedSelect` by delegating to `javaFindMember`.

This commit uses the same facility when checking that import
selectors are valid and when checking if such an import contributes
a name we are looking up.

The test cases here use `-Ypickle-java` as a straightforward
way to force typechecking all Java imports and signatures, however
the bug fix is also beneficial outside of that mode if Scala
code transitively depends on the correct interpretation of such a
Java import.

(cherry picked from ca7dc4d597e161135cb34cc2900a71530cbed89a)