Heed warning about cache invalidation after making object Serializable
Compiling the enclosed test case with `-Xdev` told me:

```
warning: !!! base trait Serializable not found in basetypes of object Person. This might indicate incorrect caching of TypeRef#parents.
```

Indeed, the `ModuleTypeRef` associated with the newly seralizable
module class had cached the base type sequence prior to the addition
of `Serializable`.

A more serious examination of how we add this parent would be
worthwhile, but for now I'll just do a bandaid job by invalidating
internal caches of the `ModuleTypeRef`.