Elim erased value types from sammy bridges
Delambdafy looks at the info of SAMs exiting erasure, so that it can
tell whether an unboxing conversion needs to be applied. If you look at
infos exiting erasure, though, you will see ErasedValueTypes, which
generally aren't valid types and under no circumstances should make it
to the backend.

It turns out that the posterasure transformer was already being applied
to the bridge def, but the symbols weren't getting the posterased type.

Fixes scala/bug#10600.