Don't lose NoWarnAttachment on ValDefs created in quasiquotes
If an anonymous function has a wilcard parameter, e.g.

```
{ _ => "hello" }
```

the `ValDef` for that parameter is given a `NoWarnAttachment` to signal
that `-Wunused:params` shouldn't warn about it being unused.

However, if we wrote the same thing in quasiquotes,

```
q"{ _ => "hello" }
```

it was triggering an unused-parameter warning.

The `NoWarnAttachment` was properly attached to the `ValDef` when it is
created, but it was being lost during reification.

The fix was to add a parameter to re-attach the attachment to the
reified tree if it was present on the original tree.

There is already precedent for doing similar things with attachments.
See e.g. https://github.com/scala/scala/blob/c14cd9433735d962da2972efc735261636b5590d/src/reflect/scala/reflect/internal/ReificationSupport.scala#L940