Xsource is Xfuture
Using `-Xsource` to revert to old behavior is not supported.

Current source level is used to mean, "version is at least,"
in order to defer changes to a future release, due to
compatibility requirements.

A few tests provided the command option in order to test
either before or after the change. However, a test using
`-Xsource:211` could be broken by other behaviors requiring
a later version. A robust test would need to compile with
the current version, and then, if current < change_version,
run again at change_version.

The `-Xsource` now requires a version argument that is
not earlier than the current version.