Reusable Instance: replace single-var with a mutable Stack.
We modify ReusableInstance, so that instead of storing a single
instance in a mutable variable, we store a mutable stack (an
ArrayBuffer) of them.

Currently, when a reentrant call to the `using` method of
ReusableInstance occurs, a new object is allocated and discarded.
Using a stack allows us to keep any number of instances around,
so that the maximum number of allocated instances is the maximum
reentrant (recursion) stack depth.