Do not rely on `ArrayIndexOOBE` to throw `IndexOOBE` in `Vector`.
This is necessary for Scala.js, where
`ArrayIndexOutOfBoundsException`s are Undefined Behavior, and
cannot be reliably caught.

I have not checked the consequences in terms of performance for the
JVM. We may have to tweak whether we should inline `checkIOOB`, or
only the test part of it but not the throwing, etc. Even the best
check may be worse than the `try..catch` on the JVM.