Inherited members have same binding level as package level
As with names introduced by package statements, base classes
can introduce definitions that shadow local definitions and
import statements.

Let inherited members have the same name-binding precedence
as package-level definitions: highest when defined in the current
compilation unit, and lowest otherwise. This means that a base
class defined elsewhere cannot introduce a name that shadows a
local definition.

The inherited name can be referred to unambiguously using `this`;
a conflicting local variable or parameter must be renamed or
aliased; a conflicting member using a `C.this` prefix.