Fix warning site for unused warnings
Even though these warnings are emitted in the typer phase, the actual
invocation is done at the end of type checking in the root context.
So we should not use the `context.warning` method, as it injects
`context.owner` as the warning site, which is always the root symbol.

(cherry picked from commit a466056367ab1cd51d5355815808ba934bc36a95)