Fixes Repl not cleaning after itself
Ref scala-dev 702

lrytz discovered that 2.13.2 nightlies do not work when called from sbt shell `console` task more than once. This is likely because JLine 3 terminal is created inside `Reader#apply` but it's never closed. So when it's invoked the second time within the same JVM we get:

```
	Suppressed: java.lang.IllegalStateException: A system terminal is already running. Make sure to use the created system Terminal on the LineReaderBuilder if you're using one or that previously created system Terminals have been correctly closed.
		at org.jline.terminal.TerminalBuilder.doBuild(TerminalBuilder.java:397)
```

To fix this I added `def close(): Unit` to `InteractiveReader` trait, which gets called after REPL is done.