Partest logs fork errors
If there is an error running a forked command, such as
a Test or javac, log it instead of letting sbt swallow
the exception. (It offers to let you run the last command
to see the stack trace.)

This helps if javac goes missing.

(cherry picked from commit f34229300dcca23438bf108b59acf893fc8d5194)