Reduce allocation cost of dead code elimination
Re-implement dead code elimination without relying on ASM's Analyzer.
Analyzers are allocation-heavy, as they allocate an array of abstract
values (a `Frame`) for every instruction in the method being analyzed.

DCE performance is important because it's enabled by default, unlike the
rest of the optimizer. The reason for that is getting cleaner bytecode.
The ASM ClassWriter doesn't eliminate dead code on the fly, but replaces
it by `throw null; nop; ...; nop;` to avoid re-allocating code array and
adjusting the offsets.

DCE now uses an abstract interpretation very similar to the existing
code in `BackendUtils.computeMaxLocalsMaxStack`. I decided to copy-paste
and adapt the implementation instead of trying to abstract over the
differences, in the interest of performance.

The new DCE implementation no longer computes the maxLocals / maxStack.
This could be done, but with the rest of the optimizer disabled these
are actually not needed. The values are needed to run an ASM Analyzer
for other optimizations, in which case they are computed when necessary.

So this PR replaces DCE by the new implementation, and it also skips the
invocation of `computeMaxLocalsMaxStack` that was previously necessary
in order to run the Analyzer.

Finally, I added a note about the interpretation of JSR/RET in
`computeMaxLocalsMaxStack` (and the new DCE), but note that they are
no longer in use since Java 7.