Fast-track emitting "eta-sam" warnings
Dotty emits the following:

    -- Warning: target/tests/EtaX/EtaX.meth1/EtaX.meth1/EtaX.meth1.01.scala:10:24 --
    10 |  val t5b: Sam1S      = meth1                   // ok, but warning
       |                        ^^^^^
       |method meth1 is eta-expanded even though Sam1S does not have the @FunctionalInterface annotation.
    1 warning found

(given the following source:)

    trait Sam1S { def apply(x: Any): Any }

    class Test {
      def meth1(x: Any) = ""
      val t5b: Sam1S      = meth1                   // ok, but warning
    }

So it makes sense for Scala 2.13 to do the same, under `-Xsource:3`.

Now it actually also emits under `-source 3.0-migration` (which used to
be called `-language:Scala2Compat`), so there's an argument that it
should be on by default in 2.13...  But let's do `-Xsource:3` at the
very least.