Avoid rare StackOverflow in diverging implicits check
Two refined types may be equal wrt `=:=` but have different symbols.
We cannot use `=:=` because it is too pessimistic (wrt wildcards),
but this property of refined types makes `allSymbols` too optimistic.

For the implicit divergence check to work, we want the invariant:
`T =:= U` implies `allSymbols(T) == allSymbols(U)` to hold.
The solution is to exclude the symbols of refined types,
which makes `allSymbols` a bit more pessimistic.

Convert `allSybols` from a recursive function to a `TypeCollector`
to avoid all the intermediate list allocations.