[2.12] Bump down baseVersion to 2.12.11
Making this change will make CrossVersion.patch work properly when using
the "merge-ly" publishes, which would now have the version
"2.12.11-bin-[sha]", which translates into the artefact suffix _2.12.11.

CrossVersion.patch is an improved version of CrossVersion.full, which
strips the `-bin-..." suffix and is what should be used for anything
that depends on the scala-compiler, such as compiler plugins.

The motivation here was to make it much, much easier (as in: actually
humanly possible) to consume the HEAD publishes of Scala when using
Scalafix (and Scalameta) in a multi-project build.