Scaladoc on AnyRef#wait
scala bug 11310

As a background, AnyRef is a fictional class for documentation only. This is achived by a special mode in scaladoc called `-doc-no-compile`, which accepts `src/library-aux`.

Internally, AnyRef is a type alias of Object (for example in definitions.AnyRefClass).
Object has 3 overloads of wait method, but `Uncompilable` mode was matching on the method name so it was returning a field named `wait` to be paired with the Scaladocs on `AnyRef#wait` methods.

This fixes that by checking both the name and the arity for `scala.AnyRef#wait`.