Fix tests
No more .flags files.

Fix some deprecation and some lint warnings.

In live.scala, change the code that reflectively extracts the state
machine object from the Runnable submitted to the execution context.
This was re-worked in 2.13.

Move AsyncStateMachine from partest to testkit. In 2.12, the file is in
partest-extras, which was merged with partest in 2.13. The JUnit project
doesn't depend on partest, but both junit and partest depend on testkit.
Even if it's just a structural dependency, it might be worth moving the
trait to `reflect` at some point.