REPL: tab completion: Cleanups in completion results
1. Completions for imports should include both term and type names.

2. Symbols that don't `exist` are filtered out.

Symptom of 1. (before the fix)

```
scala> object A { class Buh }
object A

scala> import A.B // no completions
```

A non-existing symbol is one created for a synthetic Scala classfile.
For a classfile `p/A`, the Scala compiler initially creates a module
symbol and a class symbol (before unpickling). If the class originates
in Scala, for example `package p { object A }`, the class symbol is
then marked non-existent.

Interestingly, thanks to not filtering out non-existing symbols, the
symptom of 1. was actually hidden when reading from the classpath.

Co-authored-by: Seth Tisue <seth@tisue.net>