Force annotations in TreeUnpickler and not Pattern Matching
This commit is a squash of several commits to atomically add the feature:

When indexing a scope for the first time to set up completers, add
lazy annotations to symbols. At the end of the indexing, force the annotations
and if one is scala.annotation.internal.Child[T], add the symbol of T as a
sealed child. If we are not indexing statements, such as completing the
a method and reading its parameters, force annotations as soon as each symbol
is initialised, this is ok because the owner tree for the definition is also
initialised so forward references can be safely accessed.

force sealed children after indexing scope

cache RepeatedAnnot

force all annotations early, catching more errors

Also force annotations for lookahead contexts.

It would be good to see if perhaps we can share a single annotation cache per
tasty file and force from one place.

Force annotations for each new symbol
Additionally, when sequencing statements, only force after the whole sequence is indexed.
Forceable annotations are now stored in and forced from the initial context.
Made it more lazy to print parents of a class.
If a LambdaPolyType is in a New tree, unwrap it.

force annotations at end of createMemberSymbol

only make lazy annotation if we delay its evaluation