[bug#11021]
The issue was due to incompatibility of the two features (one is parsing @hideImplicitConversion to filter out unwanted conversions, another is parsing @throws tag to add a link to a referenced exception).
The code that filters out the hidden implicit conversions tries to parse the comment in search for the @hideImplicitConversion tag. Previously this happened during ModelFactory.makeModel().
But if a comment contains a @throws tag, the code that parses this tag assumes the model is already finished (ModelFactory.modelFinished == true), since it needs to lookup the exception.
Solved by moving the implicit handling to the ModelFactory.completeModel() stage, after the model is already finished.